refactor(frontend): modularize activities screen into controller and widgets
